+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix image object-fit defaults
 - Fix useParentBlock hook to get proper parent block
 - Adjust responsive settings to be conditioned by parent block
+- Fix Table block cell colspan
 
 ## 0.9.11